Added symbol and map file to build target
Now when building the template.gbc file, the linker also outputs a .sym and .map file. These are very useful when debugging. The .sym file will contain a list of every symbol and at which address they've been placed. The .map file will contain the layout of every section and at which address they've been placed.